LMS/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django~=4.1
django-bower==5.2.0
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-nvd3==0.9.7
django-phonenumber-field==7.0.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.1
phonenumbers==8.12.55
Pillow==9.2.0
psycopg==3.1.2
psycopg2==2.9.3
pycountry==22.3.5
python-nvd3==0.14.2
python-slugify==1.1.4
reportlab==3.6.11
requests~=2.28.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.3.0
tzdata==2022.2
Unidecode==1.3.4
urllib3==1.26.12
djangorestframework~=3.13.1